requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.18.0
pandas>=2.1.0
numpy>=1.26.0
sqlalchemy>=2.0.23
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
websockets>=11.0.3
httpx>=0.25.0
jinja2>=3.1.2
itsdangerous>=2.1.2
redis>=5.0.1
pytest>=7.4.3
alembic>=1.12.1
ccxt>=4.1.19
google-generativeai>=0.3.1
python-binance>=1.0.19
email-validator>=2.1.0
